refactor: remove Service, ChitchatQASelector classes, separate config
