+++ core/agent.py
@@ -3,13 +3,11 @@ from typing import Sequence, Hashable, Any, TypeVar, Callable, List, Dict
 from itertools import compress
 import operator
 
-from core.service import Service
 from core.state_manager import StateManager, get_state
 from core.skill_manager import SkillManager
 from core.hardcode_utterances import TG_START_UTT
 from core.state_schema import Dialog, Human
 
-S = TypeVar('S', bound=Service)
 Profile = Dict[str, Any]
 
 
@@ -24,11 +22,13 @@ class Agent:
 
     def __call__(self, utterances: Sequence[str], user_telegram_ids: Sequence[Hashable],
                  user_device_types: Sequence[Any],
-                 date_times: Sequence[datetime], locations=Sequence[Any], channel_types=Sequence[str]):
+                 date_times: Sequence[datetime], locations=Sequence[Any],
+                 channel_types=Sequence[str]):
         should_reset = [utterance == TG_START_UTT for utterance in utterances]
         # here and further me stands for mongoengine
         me_users = self.state_manager.get_or_create_users(user_telegram_ids, user_device_types)
-        me_dialogs = self.state_manager.get_or_create_dialogs(me_users, locations, channel_types, should_reset)
+        me_dialogs = self.state_manager.get_or_create_dialogs(me_users, locations, channel_types,
+                                                              should_reset)
         self.state_manager.add_human_utterances(me_dialogs, utterances, date_times)
         informative_dialogs = list(compress(me_dialogs, map(operator.not_, should_reset)))
         self._update_annotations(informative_dialogs)
@@ -59,7 +59,8 @@ class Agent:
             if any(profile.values()):
                 self.state_manager.update_user_profile(me_user, profile)
 
-    def _update_utterances(self, me_dialogs: Sequence[Dialog], values: Sequence[Any], key: str) -> None:
+    def _update_utterances(self, me_dialogs: Sequence[Dialog], values: Sequence[Any],
+                           key: str) -> None:
         if values:
             utterances = [dialog.utterances[-1] for dialog in me_dialogs]
             for utt, val in zip(utterances, values):
