+++ core/connection.py
@@ -1,6 +1,5 @@
 from mongoengine import connect
 
-from core.config import HOST, PORT, DB_NAME
-
+from core.transform_config import HOST, PORT, DB_NAME
 
 state_storage = connect(host=HOST, port=PORT, db=DB_NAME)
