--- dp_requirements.txt
@@ -1,9 +1,0 @@
-git+https://github.com/deepmipt/bert.git@feat/multi_gpu
-tensorflow-gpu==1.10.0
-tensorflow-hub==0.1.1
-xgboost==0.82
-spacy==2.0.5
-gensim==2.3.0
-pybind11==2.2.3
-git+https://github.com/deepmipt/fastText.git#egg=fastText==0.8.22
-https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz#egg=en_core_web_sm==2.0.0
